moved step and continue buttons out of debug pane
