--- src/components/DebugTab.tsx
@@ -1,34 +1,0 @@
-import { Tab } from "react-bootstrap";
-import { DebugMode } from "../model/ui";
-import { useStoreActions } from "../store";
-
-type DebugTabProps = {
-  className: string | undefined;
-  eventKey: string | undefined;
-  title: string
-	mode: DebugMode;
-}
-
-export const DebugTab = (props: DebugTabProps) => {
-	const setMode = useStoreActions(actions => actions.ideLayout.setDebugMode)
-
-	return (
-		<Tab className={props.className} eventKey={props.eventKey} title={props.title} onSelect={
-      () => {
-          console.log("tab selected")
-          setMode("debug");
-        }
-      }>
-      <p className="info">Test12345</p>
-		</Tab>
-	)
-}
-
-// TODO implement
-const Debug = () => {
-  const inner =
-      <p className="info-pane-placeholder">
-        This will eventually show variables and other useful stuff.
-      </p>
-  return <div className="DebugPane">{inner}</div>;
-};
